backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
mysql-connector-python
python-dotenv
python-jose
fastapi-mail
requests
bcrypt
